Amended the GameObject::GetGridPosition() signature to return glm::ivec2, not glm::vec2

+++ src/GameObject.h
@@ -41,7 +41,7 @@ public:
 	const bool IsLoaded(void);
 	void Reposition(const glm::vec2& newPosition);
 	void Reposition(const glm::vec3& newPosition);
-	const glm::vec2 GetGridPosition(void);
+	const glm::ivec2 GetGridPosition(void);
 
 private:
 	void UpdateRotation(void);
